chore: refactor code for `evol_complexity` in separete module
